+++ packages/0x.js/CHANGELOG.md
@@ -5,6 +5,18 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
+## v1.0.6 - _September 25, 2018_
+
+    * Dependencies updated
+
+## v1.0.5 - _September 25, 2018_
+
+    * Dependencies updated
+
+## v1.0.4 - _September 21, 2018_
+
+    * Dependencies updated
+
 ## v1.0.3 - _September 19, 2018_
 
     * Drastically reduce the bundle size by removing unused parts of included contract artifacts.